+++ CHANGELOG.md
@@ -3,13 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## Unreleased
+## [unreleased]
 
 ### Fixed
 
 - Fix notification view permissions when recepient_co_group_id value exists
 - Don't show Resume button for petition status 'Denied'
 - Show View button for COU admins at petitions index page
+- Enrollment Flow approvers should not get a url of the notification sent to the enrollee.
 
 ## [3.3.3-rciam] - 2021-10-18